extract routes, models and create an __init__.py file
